+++ src/app/imports/api/alerts/server/methods.js
@@ -6,5 +6,9 @@ Meteor.methods({
   sendToSlack(message) {
     check(message, Object);
     alerts.sendToSlack(message);
+  },
+  sendToDiscord(message) {
+    check(message, Object);
+    alerts.sendToDiscord(message);
   }
 });